fix step-by-step seeking
do not rely on receiving EC_STEP_COMPLETE from the renderer, to increment
  m_nStepForwardCount
Instead, simply increment it when calling
  pFS->Step(1, NULL)

Reason: Under heavy load, the renderer might not send a EC_STEP_COMPLETE
event for *every* frame, for example if the user holds down the step key
or presses it very quickly. In this case the counter would not increment
for every frame seeked. This leads to problems when seeking backwards
after that.